+++ CHANGELOG.md
@@ -24,6 +24,7 @@
     - Implemented smooth transition for "top left" jump destination (#388)
 - Module [MIDI-CAT](./docs/MidiCat.md)
     - Implemented response curves (logarithmic/exponential) (#258)
+    - Fixed MIDI-feedback for snapped parameters (#374)
 - Module [MIDI-CAT CTX](./docs/MidiCat.md#ctx-expander)
     - Fixed broken button-handling when triggered by Parameter-mapping
 - Module [MIDI-CAT MEM](./docs/MidiCat.md#mem-expander)